Storage configuration modified
- Modified storage configuration, now is a strategy pattern.
- Now library is easy to mock and test.
- Modified documentation to fit wiith the new changes.
- Extract a function to FileUtils to be used inside a strategy
-Created the same behaviours than library had but were converted a implementation of interface.
- Created a new behaviour default which save file in cache directory to avoid overflow data saved in intern storage
