+++ lightcompressor/src/main/java/com/abedelazizshe/lightcompressorlibrary/VideoCompressor.kt
@@ -12,6 +12,7 @@ import androidx.annotation.RequiresApi
 import com.abedelazizshe.lightcompressorlibrary.compressor.Compressor.compressVideo
 import com.abedelazizshe.lightcompressorlibrary.compressor.Compressor.isRunning
 import com.abedelazizshe.lightcompressorlibrary.config.*
+import com.abedelazizshe.lightcompressorlibrary.utils.saveVideoInExternal
 import com.abedelazizshe.lightcompressorlibrary.video.Result
 import kotlinx.coroutines.*
 import java.io.File
@@ -67,21 +68,18 @@ object VideoCompressor : CoroutineScope by MainScope() {
         context: Context,
         uris: List<Uri>,
         isStreamable: Boolean = false,
-        sharedStorageConfiguration: SharedStorageConfiguration? = null,
-        appSpecificStorageConfiguration: AppSpecificStorageConfiguration? = null,
+        storageConfiguration: StorageConfiguration,
         configureWith: Configuration,
         listener: CompressionListener,
     ) {
         // Only one is allowed
-        assert(sharedStorageConfiguration == null || appSpecificStorageConfiguration == null)
         assert(configureWith.videoNames.size == uris.size)
 
         doVideoCompression(
             context,
             uris,
             isStreamable,
-            sharedStorageConfiguration,
-            appSpecificStorageConfiguration,
+            storageConfiguration,
             configureWith,
             listener,
         )
@@ -100,8 +98,7 @@ object VideoCompressor : CoroutineScope by MainScope() {
         context: Context,
         uris: List<Uri>,
         isStreamable: Boolean,
-        sharedStorageConfiguration: SharedStorageConfiguration?,
-        appSpecificStorageConfiguration: AppSpecificStorageConfiguration?,
+        storageConfiguration: StorageConfiguration,
         configuration: Configuration,
         listener: CompressionListener,
     ) {
@@ -116,8 +113,7 @@ object VideoCompressor : CoroutineScope by MainScope() {
                 val desFile = saveVideoFile(
                     context,
                     path,
-                    sharedStorageConfiguration,
-                    appSpecificStorageConfiguration,
+                    storageConfiguration,
                     isStreamable,
                     configuration.videoNames[i],
                     shouldSave = false
@@ -127,8 +123,7 @@ object VideoCompressor : CoroutineScope by MainScope() {
                     streamableFile = saveVideoFile(
                         context,
                         path,
-                        sharedStorageConfiguration,
-                        appSpecificStorageConfiguration,
+                        storageConfiguration,
                         null,
                         configuration.videoNames[i],
                         shouldSave = false
@@ -152,8 +147,7 @@ object VideoCompressor : CoroutineScope by MainScope() {
                         val savedFile = saveVideoFile(
                             context,
                             result.path,
-                            sharedStorageConfiguration,
-                            appSpecificStorageConfiguration,
+                            storageConfiguration,
                             isStreamable,
                             configuration.videoNames[i],
                             shouldSave = true
@@ -199,150 +193,22 @@ object VideoCompressor : CoroutineScope by MainScope() {
     private fun saveVideoFile(
         context: Context,
         filePath: String?,
-        sharedStorageConfiguration: SharedStorageConfiguration?,
-        appSpecificStorageConfiguration: AppSpecificStorageConfiguration?,
+        storageConfiguration: StorageConfiguration,
         isStreamable: Boolean?,
         videoName: String,
-        shouldSave: Boolean?
+        shouldSave: Boolean
     ): File? {
-        filePath?.let {
+        return filePath?.let {
             val videoFile = File(filePath)
-
-            if (sharedStorageConfiguration != null) {
-                val videoFileName = validatedFileName(
+            storageConfiguration.createFileToSave(
+                context,
+                videoFile,
+                validatedFileName(
                     videoName,
                     isStreamable
-                )
-
-                val saveLocation =
-                    when (sharedStorageConfiguration.saveAt) {
-                        SaveLocation.downloads -> {
-                            Environment.DIRECTORY_DOWNLOADS
-                        }
-
-                        SaveLocation.pictures -> {
-                            Environment.DIRECTORY_PICTURES
-                        }
-
-                        else -> {
-                            Environment.DIRECTORY_MOVIES
-                        }
-                    }
-
-                if (Build.VERSION.SDK_INT >= 29) {
-                    val fullPath =
-                        if (sharedStorageConfiguration.subFolderName != null) "$saveLocation/${sharedStorageConfiguration.subFolderName}"
-                        else saveLocation
-                    if (shouldSave == true) {
-                        saveVideoInExternal(context, videoFileName, fullPath, videoFile)
-                        File(context.filesDir, videoFileName).delete()
-                        return File("/storage/emulated/0/${fullPath}", videoFileName)
-                    }
-                    return File(context.filesDir, videoFileName)
-                } else {
-                    val savePath =
-                        Environment.getExternalStoragePublicDirectory(saveLocation)
-
-                    val fullPath =
-                        if (sharedStorageConfiguration.subFolderName != null) "$savePath/${sharedStorageConfiguration.subFolderName}"
-                        else savePath.path
-
-                    val desFile = File(fullPath, videoFileName)
-
-                    if (!desFile.exists()) {
-                        try {
-                            desFile.parentFile?.mkdirs()
-                        } catch (e: IOException) {
-                            e.printStackTrace()
-                        }
-                    }
-
-                    if (shouldSave == true) {
-                        context.openFileOutput(videoFileName, MODE_PRIVATE)
-                            .use { outputStream ->
-                                FileInputStream(videoFile).use { inputStream ->
-                                    val buf = ByteArray(4096)
-                                    while (true) {
-                                        val sz = inputStream.read(buf)
-                                        if (sz <= 0) break
-                                        outputStream.write(buf, 0, sz)
-                                    }
-
-                                }
-                            }
-
-                    }
-                    return desFile
-                }
-            } else {
-                val videoFileName = validatedFileName(
-                    videoName,
-                    isStreamable
-                )
-
-                val fullPath =
-                    if (appSpecificStorageConfiguration!!.subFolderName != null) "${appSpecificStorageConfiguration.subFolderName}/$videoFileName"
-                    else videoFileName
-
-                if (!File("${context.filesDir}/$fullPath").exists()) {
-                    File("${context.filesDir}/$fullPath").parentFile?.mkdirs()
-                }
-
-                return File(context.filesDir, fullPath)
-
-            }
-        }
-        return null
-    }
-
-
-    @RequiresApi(Build.VERSION_CODES.Q)
-    private fun saveVideoInExternal(
-        context: Context,
-        videoFileName: String,
-        saveLocation: String,
-        videoFile: File
-    ) {
-        val values = ContentValues().apply {
-
-            put(
-                MediaStore.Images.Media.DISPLAY_NAME,
-                videoFileName
+                ),
+                shouldSave
             )
-            put(MediaStore.Images.Media.MIME_TYPE, "video/mp4")
-            put(MediaStore.Images.Media.RELATIVE_PATH, saveLocation)
-            put(MediaStore.Images.Media.IS_PENDING, 1)
-        }
-
-        var collection =
-            MediaStore.Video.Media.getContentUri(MediaStore.VOLUME_EXTERNAL_PRIMARY)
-
-        if (saveLocation == Environment.DIRECTORY_DOWNLOADS) {
-            collection = MediaStore.Downloads.EXTERNAL_CONTENT_URI
-        }
-
-        val fileUri = context.contentResolver.insert(collection, values)
-
-        fileUri?.let {
-            context.contentResolver.openFileDescriptor(fileUri, "rw")
-                .use { descriptor ->
-                    descriptor?.let {
-                        FileOutputStream(descriptor.fileDescriptor).use { out ->
-                            FileInputStream(videoFile).use { inputStream ->
-                                val buf = ByteArray(4096)
-                                while (true) {
-                                    val sz = inputStream.read(buf)
-                                    if (sz <= 0) break
-                                    out.write(buf, 0, sz)
-                                }
-                            }
-                        }
-                    }
-                }
-
-            values.clear()
-            values.put(MediaStore.Video.Media.IS_PENDING, 0)
-            context.contentResolver.update(fileUri, values, null, null)
         }
     }
